:art: Refactor download cache out into separate class, as it may be useful beyond the WebStore
